# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Historial Clinico Backend",
        "description": "Aplicación backend del sistema web Historial Clnico",
        "contact": {
            "email": "dav.almeyda@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/historial-clinico-backend/public/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicio de sesión",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nroCMP",
                                    "nroDoc"
                                ],
                                "properties": {
                                    "nroCMP": {
                                        "type": "string"
                                    },
                                    "nroDoc": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nroCMP": "25480",
                                    "nroDoc": "41686893"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/getCMP": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener código de médico",
                "operationId": "getCMP",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codMedico"
                                ],
                                "properties": {
                                    "codMedico": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codMedico": "0000026144"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}